Memory leak fix!
This took a while to track down… The `_s3` property on `S3FileScanCat` was implicitly captured by the closure for waitUntil, which was called quite often. This capture would retain the current state of the `_s3` object through its capture of `this`.  If S3’s `getObject` method had been called prior to `waitUntil` was called, it would still have the JSON object it had fetched in memory.  Over time all of these objects were being retained and not let go leading to a gigantic leak.  In tracking this down, I had commented out the `putObject` call not realizing that the `getObject` call had already done the dirty deed that `waitUntil` would capture and leak.